OCR_Train/requirements.txt
click==8.0.4
cycler==0.11.0
decorator==4.4.2
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
natsort==8.1.0
networkx==2.5.1
nltk==3.6.7
numpy==1.19.5
opencv-python-headless==4.5.4.60
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.7
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
regex==2022.3.2
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.63.0
typing_extensions==4.1.1
zipp==3.6.0
